+++ src/client/sprites/Arc.ts
@@ -42,7 +42,7 @@ export default class Arc extends Sprite {
         return path;
     }
 
-    public draw(stamping?: boolean): void {
+    public draw(stamping?: true): void {
         const c = stamping ? penCtx : ctx;
 
         c.save();