+++ src/client/sprites/Button.ts
@@ -75,7 +75,7 @@ export default class Button extends Sprite {
         return path
     }
 
-    public draw(stamping?: boolean) {
+    public draw(stamping?: true) {
         const c = stamping ? penCtx : ctx;
 
         c.save();