+++ src/client/sprites/CustomPolygon.ts
@@ -86,7 +86,7 @@ export default class CustomPolygon extends Sprite {
         return path;
     }
 
-    public draw(stamping?: boolean): void {
+    public draw(stamping?: true): void {
         const c = stamping ? penCtx : ctx;
 
         c.save();