+++ src/client/sprites/Oval.ts
@@ -55,7 +55,7 @@ export default class Oval extends Sprite {
         return path;
     }
 
-    public draw(stamping?: boolean): void {
+    public draw(stamping?: true): void {
         const c = stamping ? penCtx : ctx;
 
         c.save();