+++ src/client/sprites/Rectangle.ts
@@ -54,7 +54,7 @@ export default class Rectangle extends Sprite {
         return path;
     }
 
-    public draw(stamping?: boolean): void {
+    public draw(stamping?: true): void {
         const c = stamping ? penCtx : ctx;
 
         c.save();