+++ src/client/sprites/Text.ts
@@ -64,7 +64,7 @@ export default class Text extends Sprite {
         return path
     }
 
-    public draw(stamping?: boolean) {
+    public draw(stamping?: true) {
         const c = stamping ? penCtx : ctx;
 
         c.save();